requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.2
Pillow==9.2.0
requests==2.28.1
torch==1.12.1
torchaudio==0.12.1+cpu
torchvision==0.13.1+cpu
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
